# docs/openapi.yaml
# =============================================================================
# LAB-03 - Book Management API (PUT and DELETE), Enterprise Web Application
# Development, Kabul University.
#
# POST is our own extension of the LAB-03 hand-out (which asks for GET, PUT and
# DELETE only); GET, PUT and DELETE behave exactly as the assignment requires.
#
# Base URL: http://localhost:8080/api/v3/books
#
# The catalogue is kept in memory inside BookController, so no database is
# needed: every restart restores the three initial books.
# =============================================================================
openapi: 3.0.3
info:
  title: Book Management API (LAB-03)
  version: 0.0.1-SNAPSHOT
  description: In-memory library book catalogue exposing GET, PUT and DELETE (the LAB-03 requirement) plus POST as an extension.
servers:
  - url: http://localhost:8080
    description: Local development server

paths:
  /api/v3/books:
    get:
      tags: [Books]
      summary: List all books
      operationId: getAllBooks
      responses:
        '200':
          description: The whole catalogue
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'
    post:
      tags: [Books]
      summary: Create a new book (extension, not required by LAB-03)
      operationId: createBook
      requestBody:
        required: true
        description: Title, author and number of available copies. The id is assigned by the server.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookInput'
      responses:
        '201':
          description: Book created. The created book, including its new id, is returned.
          headers:
            Location:
              description: URL of the new resource, e.g. http://localhost:8080/api/v3/books/4
              schema:
                type: string
                example: http://localhost:8080/api/v3/books/4
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'

  /api/v3/books/{bookId}:
    put:
      tags: [Books]
      summary: Update an existing book
      operationId: updateBook
      parameters:
        - $ref: '#/components/parameters/BookId'
      requestBody:
        required: true
        description: New title, author and number of available copies (the id is never taken from the body).
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookInput'
      responses:
        '200':
          description: Book updated, the updated book is returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '404':
          description: No book exists with the given id
    delete:
      tags: [Books]
      summary: Delete a book
      operationId: deleteBook
      parameters:
        - $ref: '#/components/parameters/BookId'
      responses:
        '204':
          description: Book deleted (no response body)
        '404':
          description: No book exists with the given id

components:
  parameters:
    BookId:
      name: bookId
      in: path
      required: true
      description: Identifier of the book
      schema:
        type: integer
        format: int64
        example: 2

  schemas:
    Book:
      type: object
      description: A book of the library catalogue.
      properties:
        id:
          type: integer
          format: int64
          example: 1
        title:
          type: string
          example: Java Programming
        author:
          type: string
          example: John Smith
        availableCopies:
          type: integer
          format: int32
          example: 5

    BookInput:
      type: object
      description: Request body used to create or update a book - deliberately without an id, because the client never chooses it (the server assigns it on POST and it comes from the URL on PUT).
      required: [title, author, availableCopies]
      properties:
        title:
          type: string
          example: Modern Web Development
        author:
          type: string
          example: Sara Ahmad
        availableCopies:
          type: integer
          format: int32
          example: 6